Fix all linting and TypeScript errors: React Fast Refresh compliance and type safety
- Split ThemeContext into separate files for better React Fast Refresh compliance:
  - ThemeContext.ts: exports context and types only
  - ThemeContext.tsx: exports ThemeProvider component only
  - useTheme.ts: exports hook only
- Move ThemedApp to separate component file to fix main.tsx Fast Refresh issue
- Fix TypeScript 'any' type errors in customApiData.ts with proper type definitions:
  - MockApiError details: Record<string, string | number | boolean>
  - API method parameters: Omit<MockAsset, 'id' | 'lastUpdated'> and Partial<MockAsset>
- Update imports across all affected files
- Export queryClient from main.tsx for component separation

All ESLint and TypeScript compilation errors resolved (0 errors)
