+++ samples/FluentSample/src/components/ThemedApp.tsx
@@ -1,0 +1,20 @@
+import { FluentProvider } from '@fluentui/react-components';
+import { QueryClientProvider } from '@tanstack/react-query';
+import { BrowserRouter } from 'react-router-dom';
+import { useTheme } from '../hooks/useTheme';
+import { queryClient } from '../main';
+import App from '../App';
+
+export const ThemedApp = () => {
+  const { theme } = useTheme();
+  
+  return (
+    <FluentProvider theme={theme}>
+      <QueryClientProvider client={queryClient}>
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
+      </QueryClientProvider>
+    </FluentProvider>
+  );
+};
