+++ samples/FluentSample/src/main.tsx
@@ -1,15 +1,13 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { BrowserRouter } from 'react-router-dom'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { FluentProvider } from '@fluentui/react-components'
+import { QueryClient } from '@tanstack/react-query'
 import PowerProvider from './PowerProvider.tsx'
-import { ThemeProvider, useTheme } from './contexts/ThemeContext.tsx'
+import { ThemeProvider } from './contexts/ThemeContext.tsx'
+import { ThemedApp } from './components/ThemedApp';
 import './index.css'
-import App from './App.tsx'
 
 // Create a client for TanStack Query
-const queryClient = new QueryClient({
+export const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
       staleTime: 5 * 60 * 1000, // 5 minutes
@@ -17,21 +15,6 @@ const queryClient = new QueryClient({
     },
   },
 });
-
-// Wrapper component to access theme context
-const ThemedApp = () => {
-  const { theme } = useTheme();
-  
-  return (
-    <FluentProvider theme={theme}>
-      <QueryClientProvider client={queryClient}>
-        <BrowserRouter>
-          <App />
-        </BrowserRouter>
-      </QueryClientProvider>
-    </FluentProvider>
-  );
-};
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
